backend/requirements.txt
fastapi
uvicorn
celery
redis
boto3
pillow
psycopg2-binary
sqlalchemy
python-multipart
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt
pytest
